% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{extract_unique_optimal_proxy}
\alias{extract_unique_optimal_proxy}
\title{Extract the optimal proxy}
\usage{
extract_unique_optimal_proxy(data, snp_proxy, pval_exposure, snp_exposure)
}
\arguments{
\item{data}{(data.frame) the data containing SNPs and proxies
obtained after \code{make_rank}}

\item{snp_proxy}{(character) the name of the column in \code{data} that
contains the names of the proxies (
either rs number of chromosome:position)}

\item{pval_exposure}{(character) the name of the column in \code{data} that
contains the p-value for the
association between the SNP and the
exposure}

\item{snp_exposure}{(character) the name of the column in \code{data} that
contains the SNP (either rs number or
chromosome:position)}
}
\value{
an object of class tibble
}
\description{
For each SNP, collect the most correlated proxy
}
\examples{

\dontrun{

test_df |>
  make_rank("snp_exposure", "r2_proxy") |>
  extract_unique_optimal_proxy(
  "snp_proxy", "pval_exposure", "snp_exposure"
  )

}

}
