% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{filter_optimal_proxies}
\alias{filter_optimal_proxies}
\title{Filter optimal proxies}
\usage{
filter_optimal_proxies(data, snp_proxy, pval_exposure)
}
\arguments{
\item{data}{(data.frame) the data containing SNPs and proxies obtained
after \code{make_rank}}

\item{snp_proxy}{(character) the name of the column in \code{data} that
contains the names of the proxies (
either rs number of chromosome:position)}

\item{pval_exposure}{(character) the name of the column in \code{data} that
contains the p-value for the
association between the SNP and the
exposure}
}
\value{
an object of class tibble
}
\description{
For each SNP, filter the proxies with the highest correlations
measured with r2
}
\examples{

\dontrun{

test_df |>
  make_rank("snp_exposure", "r2_proxy") |>
  filter_optimal_proxies("snp_proxy", "pval_exposure")

}

}
