% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{remove_optimal_proxies}
\alias{remove_optimal_proxies}
\title{Remove SNPs with proxies from the initial data.frame}
\usage{
remove_optimal_proxies(.from, .optimal, snp_exposure, snp_proxy)
}
\arguments{
\item{.from}{(data.frame) the initial data containing all the SNPs
and their proxies}

\item{.optimal}{(data.frame) the data.frame obtained after
\code{select_snps_and_proxies} that contains
the SNPs and their associated proxies}

\item{snp_exposure}{(character) the name of the column in \code{.from}
that contains the SNP (either rs
number or chromosome:position)}

\item{snp_proxy}{(character) the name of the column in \code{.optimal} that
contains the names of the proxies (
either rs number of chromosome:position)}
}
\value{
an object of class tibble
}
\description{
Remove from the initial data.frame the SNPs that received a proxy
(along with their proxies)
}
\examples{

\dontrun{

opt_df <- test_df |>
  make_rank("snp_exposure", "r2_proxy") |>
  extract_unique_optimal_proxy(
  "snp_proxy", "pval_exposure", "snp_exposure"
  ) |>
  select_snps_and_proxies()

final_df <- remove_optimal_proxies(
  test_df, opt_df, "snp_exposure", "snp_proxy"
)

}

}
