% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{select_snps_and_proxies}
\alias{select_snps_and_proxies}
\title{Select SNPs and proxies}
\usage{
select_snps_and_proxies(data)
}
\arguments{
\item{data}{(data.frame) the data containing SNPs and proxies obtained
after \code{extract_unique_optimal_proxy}}
}
\value{
an object of class tibble
}
\description{
Select the SNPs and their proxies along with the rank based on
the correlation
}
\examples{

\dontrun{

test_df |>
  make_rank("snp_exposure", "r2_proxy") |>
  extract_unique_optimal_proxy(
  "snp_proxy", "pval_exposure", "snp_exposure"
  ) |>
  select_snps_and_proxies()

}

}
